Convert add coauthors to TS file
#74
